Breaks up constants stories
