+++ src/helpers/getCSSVariable.tsx
@@ -1,0 +1,7 @@
+export default function getCSSVariable(name: string) {
+    let propertyValue = "";
+    propertyValue = getComputedStyle(document.documentElement).getPropertyValue(
+        name
+    );
+    return propertyValue;
+}
